requirements.txt
python-telegram-bot==12.4.2
requests==2.27.1
pytz